created an actual hash function, fixed some errors, changed tile elevation to double instead of int. (this caused me way to many problems)

+++ include/GameCore/Tile.h
@@ -7,7 +7,7 @@ enum class terrain_t { LAND, WATER, MOUNTAIN };
 struct Tile
 {
     terrain_t terrain;
-    int elevation;
+    float elevation;
 };
 
 #endif
